scrapy/requirements.txt

psycopg2-binary

scrapy

scrapy-playwright